Fix failure caused by running aqt list too early
In the previous commit, I used `aqt list-qt` to determine the version
of Qt being installed, and I ran this command before `aqtinstall` was
installed. This obviously does not work.

The trouble is, we need to know the version of Qt being installed in
two locations: 1) before we install Qt, if "cached" != "true", and
2) before setting environment variables, whether "cached" is true or not.
This solution simply runs `aqt list-qt` in both of those locations.
It is possible to refactor the script so that `aqt list-qt` only runs
once, but that's a larger change than I'm willing to make at this stage.

This change introduces a potential problem: If the Qt installation
is loaded from cache, `aqt list-qt` could return a different value than
the version that exists in cache, which would cause the Qt*_DIR
environment variables to be set improperly. Maybe at this stage, the
path should be read from the file structure instead?